FIZ MEU SITE

+++ lista1/js/exercicio011.js
@@ -3,8 +3,8 @@ function chamaFuncao()
 	var op1=document.getElementById("entrada1").value;
 	var op2=document.getElementById("entrada2").value;
 
-	op1=parseInt (op1);
-	op2=parseInt (op2);
+	op1=parseFloat (op1);
+	op2=parseFloat (op2);
 
 	var resultado=hipotenusa (op1, op2);
 	alert("A hipotenusa do triangulo retangulo de cateto " + op1 + " e " + op2 + " é " + resultado);
